+++ packages/twenty-server/src/modules/calendar/common/types/calendar-event.ts
