--- packages/twenty-server/src/modules/calendar/utils/google-calendar-search-filter.util.ts
@@ -1,9 +1,0 @@
-export const googleCalendarSearchFilterExcludeEmails = (
-  emails: string[],
-): string | undefined => {
-  if (emails.length === 0) {
-    return undefined;
-  }
-
-  return `-(${emails.join(', ')})`;
-};
